Created a var for background image in In SimpleCarousel.module.css
Refactor SimpleCarousel component for cleaning

Refactor Slide component and made the correct types

Created "data" file and put there the slidesData code (slides.ts) and also created its correctly types
