+++ src/data/slides.ts
@@ -1,0 +1,23 @@
+import { type SlidesData } from "../types";
+
+// Carrusel de información
+export const slidesData : SlidesData[] = [
+  {
+    id: 1,
+    title: "Descubre la forma de hacer más",
+    text: "Administra tu inventario en tiempo real, optimiza tus recursos y ten el control total de tus productos desde cualquier lugar.",
+    className: "img-bg-descubre",
+  },
+  {
+    id: 2,
+    title: "Transforma tu manera de gestionar productos.",
+    text: "Optimiza y controla tu inventario como nunca antes. Gestionar productos es más fácil, rápido y eficiente.",
+    className: "img-bg-transforma",
+  },
+  {
+    id: 3,
+    title: "Todo lo que necesitas, en un solo lugar.",
+    text: "Controla el stock en tiempo real, registra entradas y salidas, y genera reportes detallados.",
+    className: "img-bg-optimiza",
+  },
+];
