+++ src/types/index.tsx
@@ -1,6 +1,6 @@
-export type SlideProps = {
-  title: string;
-  text: string;
-  className: string;
-  visible: boolean;
-};
+export type SlidesData = {
+  id: number
+  title: string
+  text: string
+  className: string
+}
